+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added configurable From: email address for log files
 - Fixed missing path variable that kept ```deploy ---automate``` from running correctly
 - Emails are now sent using Sendmail
 - Slack webhook URL is no longer displayed in logs